Add Lucy Warwick Ching
This is not a good cut out job. The hair is rather pointy and there are
bits of background left in at the top. I've asked for a better cut out
but Lucy's first column is tomorrow so in the interests of having
something there at all I'm adding this, and we'll fix it later.